requirements.txt
Django>=5.0,<5.2
djangorestframework>=3.15,<3.16
django-cors-headers>=4.4,<5
requests>=2.31,<3
python-dotenv>=1.0,<2
